CAR-5804: write stories in .tsx
-- retyped some stories to tsx
