+++ src/stories/checkbox/buttonStyleCheckbox.stories.tsx
@@ -1,4 +1,4 @@
-import React from "react"
+import React, { FC } from "react"
 
 import { action } from "@storybook/addon-actions"
 
@@ -28,7 +28,13 @@ export default {
   },
 }
 
-const Template = (args) => {
+interface Props {
+  storyName: string
+  label: JSX.Element
+  name: string
+}
+
+const Template: FC<Props> = ({ ...args } : Props) => {
   return (
     <StoryContainer
       title={args.storyName}
