+++ src/stories/input/floatingLabel.stories.tsx
@@ -1,4 +1,4 @@
-import React from "react"
+import React, { FC } from "react"
 import { action } from "@storybook/addon-actions"
 
 import StoryContainer from "../storyContainer"
@@ -6,19 +6,28 @@ import Input from "../../components/input/index"
 
 const onBlur = () => action("onBlur")
 const onChange = () => action("onChange")
-
-const setValue = (value) => ("value", value)
+const setValue = (value: string) => ("value", value)
 
 export default {
-  title: "Input/Label",
+  title: "Input/Floating Label",
   component: Input,
   args: {
-    storyName: "Input / Label",
-    labelText: "Label",
+    storyName: "Input / Floating Label",
   },
 }
 
-const Template = (args) => {
+interface Props {
+  storyName: string
+  label: string
+  name: string
+  value: string
+  mode: "text" | "numeric" | "decimal" | "tel" | "email"
+  onChange: <T extends { target: { name: string; value: string | number } }>(
+    e: T
+  ) => void
+}
+
+const Template: FC<Props> = ({ ...args } : Props) => {
   return (
     <StoryContainer
       title={args.storyName}
@@ -31,12 +40,16 @@ export const Standard = Template.bind({})
 Standard.args = {
   storyName: "Standard",
   value: "",
+  labelText: "Label",
+  floatingLabel: true,
 }
 
 export const WithErrorMessage = Template.bind({})
 WithErrorMessage.args = {
   storyName: "With Error Message",
   value: "",
+  labelText: "Label",
+  floatingLabel: true,
   error: "Error message",
 }
 
@@ -45,7 +58,8 @@ WithClearButton.args = {
   storyName: "With clear button",
   value: "text",
   hasClearButton: true,
-  placeholder: "Type something",
+  labelText: "Label",
+  floatingLabel: true,
   onChange: (e) => {
     setValue(e.target.value)
     onChange()(e)
@@ -56,6 +70,7 @@ WithClearButton.args = {
 export const WithPlaceholder = Template.bind({})
 WithPlaceholder.args = {
   storyName: "With placeholder",
+  value: "",
   placeholder: "Placeholder",
 }
 
@@ -64,26 +79,6 @@ WithHint.args = {
   storyName: "With hint",
   value: "",
   hint: "Hint text",
-}
-
-export const WithRequiredIndicator = Template.bind({})
-WithRequiredIndicator.args = {
-  storyName: "With required indicator",
-  value: "",
-  required: true,
-}
-
-export const WithPopup = Template.bind({})
-WithPopup.args = {
-  storyName: "With Popup",
-  value: "",
-  renderLabelPopup: () => <div>Popup Content</div>,
-}
-
-export const RequiredWithPopup = Template.bind({})
-RequiredWithPopup.args = {
-  storyName: "Required with popup",
-  value: "",
-  required: true,
-  renderLabelPopup: () => <div>Popup Content</div>,
+  labelText: "Label",
+  floatingLabel: true,
 }
